[Video] Extract stream creation from VideoStreamSelectHelper
